requirements.txt
# 国内ETF智能分析系统 - Python依赖

# 数据获取
akshare>=1.14.0          # A股/ETF数据源（免费开源）
pandas>=2.0.0
numpy>=1.24.0

# HTTP客户端（LLM API + 推送）
httpx>=0.27.0
aiohttp>=3.9.0

# 配置管理
python-dotenv>=1.0.0

# Web服务
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0

# 定时任务
APScheduler>=3.10.0
pytz>=2024.1

# 可选：Tushare数据源（需申请token）
# tushare>=1.4.17
